refactor: SetTitle.tsx component 분리
